RET-172 файл Dropdown расспличен на компоненты
